1.if user login first time, write a sample training data 2.create useContext for user login status at App.js
