+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "fitness2-d4aaf"
+  }
+}
