+++ src/pages/Training/AllActions.js
@@ -1,0 +1,244 @@
+const all = [
+  {
+    actionName: '槓鈴肩推',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E6%A7%93%E9%88%B4%E8%82%A9%E6%8E%A8.mp4?alt=media&token=b7a4b84a-8320-40f3-9889-9814fae3fdf1',
+  },
+  {
+    actionName: '坐姿肩推',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E5%9D%90%E5%A7%BF%E8%82%A9%E6%8E%A8.mp4?alt=media&token=b5b90252-7e6c-4d63-b26f-f65dee987d31',
+  },
+  {
+    actionName: '側平舉',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E5%81%B4%E5%B9%B3%E8%88%89.mp4?alt=media&token=f9e62e29-dca8-4f04-a988-fe016dea997f',
+  },
+  {
+    actionName: '蝴蝶飛鳥機',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E8%9D%B4%E8%9D%B6%E9%A3%9B%E9%B3%A5%E6%A9%9F.mp4?alt=media&token=05174b90-6af6-46be-84cf-8eeee5d25f50',
+  },
+  {
+    actionName: '碎顱者',
+    bodyPart: '手臂',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Farm%2F%E7%A2%8E%E9%A1%B1%E8%80%85.mp4?alt=media&token=223a5d6c-0f29-4cc2-a4f7-c302da4a60bbs',
+  },
+  {
+    actionName: '二頭彎舉訓練機',
+    bodyPart: '手臂',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Farm%2F%E4%BA%8C%E9%A0%AD%E5%BD%8E%E8%88%89%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=4ebc7c36-f727-4b7b-bb27-3533e499087c',
+  },
+  {
+    actionName: '三頭肌伸展訓練機',
+    bodyPart: '手臂',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Farm%2F%E4%B8%89%E9%A0%AD%E8%82%8C%E4%BC%B8%E5%B1%95%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=aff44ade-7abd-4c85-80ae-6f99a6be6aef',
+  },
+  {
+    actionName: '槓鈴臥推',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E6%A7%93%E9%88%B4%E8%87%A5%E6%8E%A8.mp4?alt=media&token=d5164b79-2be9-490b-8865-4220c87c347b',
+  },
+  {
+    actionName: '滑輪機',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E6%BB%91%E8%BC%AA%E6%A9%9F.mp4?alt=media&token=b3095e8c-f679-4735-bc99-d51ef0e0e19a',
+  },
+  {
+    actionName: '蝴蝶夾胸機',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E8%9D%B4%E8%9D%B6%E5%A4%BE%E8%83%B8%E6%A9%9F.mp4?alt=media&token=6b6e0335-fca6-4ed3-8dd2-7338fa5d0f12',
+  },
+  {
+    actionName: '胸推',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E8%83%B8%E6%8E%A8.mp4?alt=media&token=3e83396b-6247-478b-8d70-0d685127b97c',
+  },
+  {
+    actionName: '上斜胸推',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E4%B8%8A%E6%96%9C%E8%83%B8%E6%8E%A8.mp4?alt=media&token=698614fd-ba2f-4026-a87e-019a515e583f',
+  },
+  {
+    actionName: '槓鈴划船',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%A7%93%E9%88%B4%E5%88%92%E8%88%B9.mp4?alt=media&token=eec0d322-aa7f-4639-8621-ce2163dbcc86',
+  },
+  {
+    actionName: '槓鈴俯身划船',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%A7%93%E9%88%B4%E4%BF%AF%E8%BA%AB%E5%88%92%E8%88%B9.mp4?alt=media&token=789856d2-0807-4fe3-a996-cc19870c222b',
+  },
+  {
+    actionName: '啞鈴划船',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E5%95%9E%E9%88%B4%E5%88%92%E8%88%B9.mp4?alt=media&token=6b99c720-cd76-446c-acca-d796f0b7685e',
+  },
+  {
+    actionName: '水平引體',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%B0%B4%E5%B9%B3%E5%BC%95%E9%AB%94.mp4?alt=media&token=f5a596de-5537-4e7d-ac71-8970e3d97813',
+  },
+  {
+    actionName: '低位划船機',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E4%BD%8E%E4%BD%8D%E5%88%92%E8%88%B9%E6%A9%9F.mp4?alt=media&token=e7f4ae60-756b-47e8-9b4e-8176d9e298f1',
+  },
+  {
+    actionName: '滑輪下拉機',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%BB%91%E8%BC%AA%E4%B8%8B%E6%8B%89%E6%A9%9F.mp4?alt=media&token=d48fc8c4-9b78-472b-9496-ee0d9a52227c',
+  },
+  {
+    actionName: '背部伸張訓練機',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E8%83%8C%E9%83%A8%E4%BC%B8%E5%BC%B5%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=5102e5e0-95b0-448c-994c-22526871bab4',
+  },
+  {
+    actionName: '臀推',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E8%87%80%E6%8E%A8.mp4?alt=media&token=843d6b57-9bdf-4486-8c5d-d5b261a7a464',
+  },
+  {
+    actionName: '雙腿伸屈',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E9%9B%99%E8%85%BF%E4%BC%B8%E5%B1%88.mp4?alt=media&token=8b8674a4-91ed-4e1c-a08b-130dcdd7b00e',
+  },
+  {
+    actionName: '雙腿彎舉',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E9%9B%99%E8%85%BF%E5%BD%8E%E8%88%89.mp4?alt=media&token=929da96d-f385-4ebe-8f91-d8086353fcc0',
+  },
+  {
+    actionName: '深蹲',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E6%B7%B1%E8%B9%B2.mp4?alt=media&token=61f35d20-558f-4ee0-940c-8588e0b5fd17',
+  },
+  {
+    actionName: '腿推機',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E8%85%BF%E6%8E%A8%E6%A9%9F.mp4?alt=media&token=05fe219d-1605-45b4-8035-f4a5a3bde281',
+  },
+  {
+    actionName: '夾腿機',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E5%A4%BE%E8%85%BF%E6%A9%9F.mp4?alt=media&token=57991aec-b451-473b-80fd-8c2ba521878e',
+  },
+  {
+    actionName: '坐姿捲腹機',
+    bodyPart: '核心',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fcore%2F%E5%9D%90%E5%A7%BF%E6%8D%B2%E8%85%B9%E6%A9%9F.mp4?alt=media&token=eb10df30-cd9d-44f8-8b44-dcd502d701c4',
+  },
+  {
+    actionName: '轉體機',
+    bodyPart: '核心',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fcore%2F%E8%BD%89%E9%AB%94%E6%A9%9F.mp4?alt=media&token=048b56fb-7fef-4c77-8a55-2b72363a02bf',
+  },
+  {
+    actionName: '腹部訓練機',
+    bodyPart: '核心',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fcore%2F%E8%85%B9%E9%83%A8%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=f2d8dbc9-f8c6-44e9-ae55-0c5068d94616',
+  },
+  {
+    actionName: '背部伸展',
+    bodyPart: '核心',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fcore%2F%E8%83%8C%E9%83%A8%E4%BC%B8%E5%B1%95.mp4?alt=media&token=8d437cf2-07eb-4fce-945a-e825284ccd6e',
+  },
+  {
+    actionName: '側棒式',
+    bodyPart: '核心',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fcore%2F%E5%81%B4%E6%A3%92%E5%BC%8F.mp4?alt=media&token=fa73e2f8-8f37-4c33-ba34-aef75dfe8b01',
+  },
+];
+
+export default all;
