+++ src/pages/Training/ArmActions.js
@@ -1,0 +1,28 @@
+const arm = [
+  {
+    actionName: '碎顱者',
+    bodyPart: '手臂',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Farm%2F%E7%A2%8E%E9%A1%B1%E8%80%85.mp4?alt=media&token=223a5d6c-0f29-4cc2-a4f7-c302da4a60bbs',
+  },
+  {
+    actionName: '二頭彎舉訓練機',
+    bodyPart: '手臂',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Farm%2F%E4%BA%8C%E9%A0%AD%E5%BD%8E%E8%88%89%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=4ebc7c36-f727-4b7b-bb27-3533e499087c',
+  },
+  {
+    actionName: '三頭肌伸展訓練機',
+    bodyPart: '手臂',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Farm%2F%E4%B8%89%E9%A0%AD%E8%82%8C%E4%BC%B8%E5%B1%95%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=aff44ade-7abd-4c85-80ae-6f99a6be6aef',
+  },
+];
+
+export default arm;
