+++ src/pages/Training/BackActions.js
@@ -1,0 +1,60 @@
+const back = [
+  {
+    actionName: '槓鈴划船',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%A7%93%E9%88%B4%E5%88%92%E8%88%B9.mp4?alt=media&token=eec0d322-aa7f-4639-8621-ce2163dbcc86',
+  },
+  {
+    actionName: '槓鈴俯身划船',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%A7%93%E9%88%B4%E4%BF%AF%E8%BA%AB%E5%88%92%E8%88%B9.mp4?alt=media&token=789856d2-0807-4fe3-a996-cc19870c222b',
+  },
+  {
+    actionName: '啞鈴划船',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E5%95%9E%E9%88%B4%E5%88%92%E8%88%B9.mp4?alt=media&token=6b99c720-cd76-446c-acca-d796f0b7685e',
+  },
+  {
+    actionName: '水平引體',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%B0%B4%E5%B9%B3%E5%BC%95%E9%AB%94.mp4?alt=media&token=f5a596de-5537-4e7d-ac71-8970e3d97813',
+  },
+  {
+    actionName: '低位划船機',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E4%BD%8E%E4%BD%8D%E5%88%92%E8%88%B9%E6%A9%9F.mp4?alt=media&token=e7f4ae60-756b-47e8-9b4e-8176d9e298f1',
+  },
+  {
+    actionName: '滑輪下拉機',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%BB%91%E8%BC%AA%E4%B8%8B%E6%8B%89%E6%A9%9F.mp4?alt=media&token=d48fc8c4-9b78-472b-9496-ee0d9a52227c',
+  },
+  {
+    actionName: '背部伸張訓練機',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E8%83%8C%E9%83%A8%E4%BC%B8%E5%BC%B5%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=5102e5e0-95b0-448c-994c-22526871bab4',
+  },
+];
+
+export default back;
