+++ src/pages/Training/ButtLegActions.js
@@ -1,0 +1,51 @@
+const buttLeg = [
+  {
+    actionName: '臀推',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E8%87%80%E6%8E%A8.mp4?alt=media&token=843d6b57-9bdf-4486-8c5d-d5b261a7a464',
+  },
+  {
+    actionName: '雙腿伸屈',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E9%9B%99%E8%85%BF%E4%BC%B8%E5%B1%88.mp4?alt=media&token=8b8674a4-91ed-4e1c-a08b-130dcdd7b00e',
+  },
+  {
+    actionName: '雙腿彎舉',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E9%9B%99%E8%85%BF%E5%BD%8E%E8%88%89.mp4?alt=media&token=929da96d-f385-4ebe-8f91-d8086353fcc0',
+  },
+  {
+    actionName: '深蹲',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E6%B7%B1%E8%B9%B2.mp4?alt=media&token=61f35d20-558f-4ee0-940c-8588e0b5fd17',
+  },
+  {
+    actionName: '腿推機',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E8%85%BF%E6%8E%A8%E6%A9%9F.mp4?alt=media&token=05fe219d-1605-45b4-8035-f4a5a3bde281',
+  },
+  {
+    actionName: '夾腿機',
+    bodyPart: '臀腿',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2FbuttLeg%2F%E5%A4%BE%E8%85%BF%E6%A9%9F.mp4?alt=media&token=57991aec-b451-473b-80fd-8c2ba521878e',
+  },
+];
+export default buttLeg;
