+++ src/pages/Training/ChestActions.js
@@ -1,0 +1,44 @@
+const chest = [
+  {
+    actionName: '槓鈴臥推',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E6%A7%93%E9%88%B4%E8%87%A5%E6%8E%A8.mp4?alt=media&token=d5164b79-2be9-490b-8865-4220c87c347b',
+  },
+  {
+    actionName: '滑輪機',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E6%BB%91%E8%BC%AA%E6%A9%9F.mp4?alt=media&token=b3095e8c-f679-4735-bc99-d51ef0e0e19a',
+  },
+  {
+    actionName: '蝴蝶夾胸機',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E8%9D%B4%E8%9D%B6%E5%A4%BE%E8%83%B8%E6%A9%9F.mp4?alt=media&token=6b6e0335-fca6-4ed3-8dd2-7338fa5d0f12',
+  },
+  {
+    actionName: '胸推',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E8%83%B8%E6%8E%A8.mp4?alt=media&token=3e83396b-6247-478b-8d70-0d685127b97c',
+  },
+  {
+    actionName: '上斜胸推',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E4%B8%8A%E6%96%9C%E8%83%B8%E6%8E%A8.mp4?alt=media&token=698614fd-ba2f-4026-a87e-019a515e583f',
+  },
+];
+
+export default chest;
