+++ src/pages/Training/CoreActions.js
@@ -1,0 +1,44 @@
+const core = [
+  {
+    actionName: '坐姿捲腹機',
+    bodyPart: '核心',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fcore%2F%E5%9D%90%E5%A7%BF%E6%8D%B2%E8%85%B9%E6%A9%9F.mp4?alt=media&token=eb10df30-cd9d-44f8-8b44-dcd502d701c4',
+  },
+  {
+    actionName: '轉體機',
+    bodyPart: '核心',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fcore%2F%E8%BD%89%E9%AB%94%E6%A9%9F.mp4?alt=media&token=048b56fb-7fef-4c77-8a55-2b72363a02bf',
+  },
+  {
+    actionName: '腹部訓練機',
+    bodyPart: '核心',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fcore%2F%E8%85%B9%E9%83%A8%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=f2d8dbc9-f8c6-44e9-ae55-0c5068d94616',
+  },
+  {
+    actionName: '背部伸展',
+    bodyPart: '核心',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fcore%2F%E8%83%8C%E9%83%A8%E4%BC%B8%E5%B1%95.mp4?alt=media&token=8d437cf2-07eb-4fce-945a-e825284ccd6e',
+  },
+  {
+    actionName: '側棒式',
+    bodyPart: '核心',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fcore%2F%E5%81%B4%E6%A3%92%E5%BC%8F.mp4?alt=media&token=fa73e2f8-8f37-4c33-ba34-aef75dfe8b01',
+  },
+];
+
+export default core;
