+++ src/pages/Training/ShoulderActions.js
@@ -1,0 +1,36 @@
+const shoulder = [
+  {
+    actionName: '槓鈴肩推',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E6%A7%93%E9%88%B4%E8%82%A9%E6%8E%A8.mp4?alt=media&token=b7a4b84a-8320-40f3-9889-9814fae3fdf1',
+  },
+  {
+    actionName: '坐姿肩推',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E5%9D%90%E5%A7%BF%E8%82%A9%E6%8E%A8.mp4?alt=media&token=b5b90252-7e6c-4d63-b26f-f65dee987d31',
+  },
+  {
+    actionName: '側平舉',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E5%81%B4%E5%B9%B3%E8%88%89.mp4?alt=media&token=f9e62e29-dca8-4f04-a988-fe016dea997f',
+  },
+  {
+    actionName: '蝴蝶飛鳥機',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E8%9D%B4%E8%9D%B6%E9%A3%9B%E9%B3%A5%E6%A9%9F.mp4?alt=media&token=05174b90-6af6-46be-84cf-8eeee5d25f50',
+  },
+];
+
+export default shoulder;
