+++ src/pages/Training/UpperBodyActions.js
@@ -1,0 +1,156 @@
+const upperBody = [
+  {
+    actionName: '槓鈴肩推',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E6%A7%93%E9%88%B4%E8%82%A9%E6%8E%A8.mp4?alt=media&token=b7a4b84a-8320-40f3-9889-9814fae3fdf1',
+  },
+  {
+    actionName: '坐姿肩推',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E5%9D%90%E5%A7%BF%E8%82%A9%E6%8E%A8.mp4?alt=media&token=b5b90252-7e6c-4d63-b26f-f65dee987d31',
+  },
+  {
+    actionName: '側平舉',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E5%81%B4%E5%B9%B3%E8%88%89.mp4?alt=media&token=f9e62e29-dca8-4f04-a988-fe016dea997f',
+  },
+  {
+    actionName: '蝴蝶飛鳥機',
+    bodyPart: '肩',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fshoulder%2F%E8%9D%B4%E8%9D%B6%E9%A3%9B%E9%B3%A5%E6%A9%9F.mp4?alt=media&token=05174b90-6af6-46be-84cf-8eeee5d25f50',
+  },
+  {
+    actionName: '碎顱者',
+    bodyPart: '手臂',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Farm%2F%E7%A2%8E%E9%A1%B1%E8%80%85.mp4?alt=media&token=223a5d6c-0f29-4cc2-a4f7-c302da4a60bbs',
+  },
+  {
+    actionName: '二頭彎舉訓練機',
+    bodyPart: '手臂',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Farm%2F%E4%BA%8C%E9%A0%AD%E5%BD%8E%E8%88%89%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=4ebc7c36-f727-4b7b-bb27-3533e499087c',
+  },
+  {
+    actionName: '三頭肌伸展訓練機',
+    bodyPart: '手臂',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Farm%2F%E4%B8%89%E9%A0%AD%E8%82%8C%E4%BC%B8%E5%B1%95%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=aff44ade-7abd-4c85-80ae-6f99a6be6aef',
+  },
+  {
+    actionName: '槓鈴臥推',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E6%A7%93%E9%88%B4%E8%87%A5%E6%8E%A8.mp4?alt=media&token=d5164b79-2be9-490b-8865-4220c87c347b',
+  },
+  {
+    actionName: '滑輪機',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E6%BB%91%E8%BC%AA%E6%A9%9F.mp4?alt=media&token=b3095e8c-f679-4735-bc99-d51ef0e0e19a',
+  },
+  {
+    actionName: '蝴蝶夾胸機',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E8%9D%B4%E8%9D%B6%E5%A4%BE%E8%83%B8%E6%A9%9F.mp4?alt=media&token=6b6e0335-fca6-4ed3-8dd2-7338fa5d0f12',
+  },
+  {
+    actionName: '胸推',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E8%83%B8%E6%8E%A8.mp4?alt=media&token=3e83396b-6247-478b-8d70-0d685127b97c',
+  },
+  {
+    actionName: '上斜胸推',
+    bodyPart: '胸',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fchest%2F%E4%B8%8A%E6%96%9C%E8%83%B8%E6%8E%A8.mp4?alt=media&token=698614fd-ba2f-4026-a87e-019a515e583f',
+  },
+  {
+    actionName: '槓鈴划船',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%A7%93%E9%88%B4%E5%88%92%E8%88%B9.mp4?alt=media&token=eec0d322-aa7f-4639-8621-ce2163dbcc86',
+  },
+  {
+    actionName: '槓鈴俯身划船',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%A7%93%E9%88%B4%E4%BF%AF%E8%BA%AB%E5%88%92%E8%88%B9.mp4?alt=media&token=789856d2-0807-4fe3-a996-cc19870c222b',
+  },
+  {
+    actionName: '啞鈴划船',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E5%95%9E%E9%88%B4%E5%88%92%E8%88%B9.mp4?alt=media&token=6b99c720-cd76-446c-acca-d796f0b7685e',
+  },
+  {
+    actionName: '水平引體',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%B0%B4%E5%B9%B3%E5%BC%95%E9%AB%94.mp4?alt=media&token=f5a596de-5537-4e7d-ac71-8970e3d97813',
+  },
+  {
+    actionName: '低位划船機',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E4%BD%8E%E4%BD%8D%E5%88%92%E8%88%B9%E6%A9%9F.mp4?alt=media&token=e7f4ae60-756b-47e8-9b4e-8176d9e298f1',
+  },
+  {
+    actionName: '滑輪下拉機',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E6%BB%91%E8%BC%AA%E4%B8%8B%E6%8B%89%E6%A9%9F.mp4?alt=media&token=d48fc8c4-9b78-472b-9496-ee0d9a52227c',
+  },
+  {
+    actionName: '背部伸張訓練機',
+    bodyPart: '背',
+    weight: 0,
+    times: 0,
+    videoURL:
+      'https://firebasestorage.googleapis.com/v0/b/fitness2-d4aaf.appspot.com/o/videos%2Fback%2F%E8%83%8C%E9%83%A8%E4%BC%B8%E5%BC%B5%E8%A8%93%E7%B7%B4%E6%A9%9F.mp4?alt=media&token=5102e5e0-95b0-448c-994c-22526871bab4',
+  },
+];
+
+export default upperBody;
